+++ backend/sparrow/interface/converter.py
@@ -8,11 +8,11 @@ from marshmallow_sqlalchemy.fields import Related, RelatedList
 
 import geoalchemy2 as geo
 from sqlalchemy.orm import RelationshipProperty
-from sqlalchemy.types import Integer, Numeric
+from sqlalchemy.types import Integer, Numeric, DateTime
 from sqlalchemy.dialects import postgresql
 
 from ..database.mapper.util import trim_postfix
-from .fields import Geometry, Enum, JSON, SmartNested, UUID, PassThroughRelated
+from .fields import Geometry, Enum, JSON, SmartNested, UUID, PassThroughRelated, DateTimeExt
 from .util import to_schema_name
 
 from ..logs import get_logger
@@ -91,6 +91,7 @@ class SparrowConverter(ModelConverter):
                 postgresql.JSON: JSON,
                 postgresql.JSONB: JSON,
                 postgresql.UUID: UUID,
+                DateTime: DateTimeExt
             }.items()
         )
     )